Updates to failed DB access and Elastic Pool
Script has been updated to support "access denied or unavailable" DBs.  In addition, due to how the ingestion API inserts data into Log Analytics, if an elastic pool is not present for any DB, the field is not created.  Script has been updated to support a value of "none" if the elastic pool does not exist.  This will be relevant when we provide an OMS view for this solution.